requirements.txt
coloredlogs
requests
sh
pyyaml